WIP things

+++ Configs/Config.cs
@@ -13,7 +13,7 @@ namespace UIURescueSquad.Configs
         public bool Debug { get; set; } = false;
 
         [Description("How many seconds before a spawnwave occurs should it calculate the spawn chance. **NOTE** this value must be lower than 18")]
-        public int SpawnWaveCalculation { get; set; } = 10;
+        public float SpawnWaveCalculation { get; set; } = 10;
 
         [Description("Options for UIU spawn:")]
         public SpawnManager SpawnManager { get; private set; } = new SpawnManager();
